Use MUI for left navigation drawer
