+++ package.json
@@ -21,6 +21,11 @@
     "webpack-dev-server": "4.13.3"
   },
   "dependencies": {
+    "@emotion/react": "11.11.1",
+    "@emotion/styled": "11.11.0",
+    "@fontsource/roboto": "5.0.4",
+    "@mui/icons-material": "5.11.16",
+    "@mui/material": "5.13.6",
     "@react-spring/web": "9.7.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
